Offline AI: fix landscape on very wide screens.
The 'Create game' button does not work in landscape mode on very wide
screens (see issue #250 for more info). Besides, it can accidentally be
pressed and, without any confirmation dialog, the current game is lost.
It's better to just have the ellipsis button, like in landscape on small
screens.

This fixes issue #250.