Extend watch-scripts gulp task to rebundle for the platforms directories.
Extend the watch-scripts gulp task to rebundle the version of app.js
(uglified) that goes in platforms/{android,ios}/www.

For example, this means that if a developer is using Xcode to run the
app, they can use "npm run watch-stage" and any changes in the source
code will get rebundled into platforms/ios/www/app.js automatically.
The developer still needs to restart the app inside Xcode (to copy
the files into the simulator and restart the app) but this is a faster
turnaround than "npm run prepare-ios-stage".

I have implemented this as a glob for platforms/{android,ios}/www.
I don't actually know if this is the correct workflow for android
developers using IDEs, but I presume that it is.  If not, we can
change the glob to just cover platforms/ios.