Allow building without firebase-messaging.
This commit provides a way to build lichobile for Android without
support for push notifications, dropping the dependency on Firebase.
This is a required step towards F-Droid inclusion.

The main idea is to expose a `patch-nonfree` script that will:

- replace the push notification implementation with a stub
- remove any direct or indirect dependency to firebase (there was an
  indirect one in the @capacitor/push-notifications plugin, and a direct
  one in the app/build.gradle)
- change the applicationId so that it doesn't gets confused with the
  non-free one

The implementation is not as elegant as I'd have liked because I haven't
found a magic way to enable or disable dependencies for specific product
flavors with capacitor ([as the Android gradle plugin allows](
https://developer.android.com/studio/build/build-variants#product-flavors)).
I'm far from a capacitor expert though, so there may be better ways.

As I had to patch the JS sources anyways, I chose to not rely on
Android's product flavors at all to prevent any confusion that would
lead to build a non-free version as a free APK and vice-versa. I think
it is pretty straightforward, not too intrusive, and that it should be
manageable maintenance-wise.

The remaining changes in the JS code are needed to remove any mention of
"notifications" in the UI when using the push stub. We don't expect
notifications to work on the free version.

Fixes https://github.com/lichess-org/lichobile/issues/1121

(well, technically this doesn't "remove" the dependency to Firebase as
the main build flavor still relies on it, but I think getting rid of
that is lengthier project as lila itself depends on it)

Issues related to F-Droid inclusion:

- https://github.com/lichess-org/lichobile/issues/278
- https://github.com/lichess-org/lichobile/issues/409
- https://github.com/lichess-org/lichobile/issues/927
- https://github.com/lichess-org/lichobile/issues/1331