Add setting to disable back button during game
This is very useful on Android 12 without dedicated buttons, as swiping
from the edge of the screen generates a back button press event. Under
time pressure moving pieces on the edge of the board may easily cause
navigation out of the game and greatly affect the experience. I've lost
a dozen blitz games like this.

With the setting enabled, a back button press simply does nothing which
is strictly better. Even better would be to disable the swiping gestures
during game altogether, but I don't know of a way to do that.