Fix sorting and filtering studies
Previously, whenever you would use one of the filter or sort controls on the study page, it would reset the other one. That meant that it was impossible to view your own studies sorted by recently updated for example. This change preserves the filter when you change the sort and vice versa. Here's a video showing that the change works: https://d.rayll.com/i/xhN0OG/XRTuSzRk1a